backend_server/requirements.txt
flask
flask-cors
feedparser
chromadb
openai
python-dotenv
requests
pypdf
edge-tts
pymongo
cloudinary
bcrypt
PyJWT
google-api-python-client
matplotlib
pandas
numpy
